+++ frontend/src/components/EventBadge.tsx
