+++ frontend/src/components/EventHeader.tsx
