+++ frontend/src/components/Statistic.tsx
