+++ frontend/src/routes/events/OverviewTabs/ChallengesTab/ChallengeCard.tsx
